requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.1
textstat==0.7.3